compute_metrics: one function for each metric (unoptimized at the moment) consolid_metrics: created a variable for the name sof quality metrics (could refer to constants in the future) constants: changed name of output directory and order of the metrics fa2util.pxd: included the dimension parameter in the relevant functions fa2util.py: reworked some functions so it can accept 2 or 3 dimensions and do calculations accordingly forceatlas2.py: included support for a dimension parameter metrics_gd.py: added margin as a parameter for edge edge occlusion pivot_mds.py: slightly adjusted a check for if the gtds needed to be calculated requirements_new.txt: included cython and tqdm run_fa2.py: included numpy import, dimension as parameter, assertion, coordinate return and barneshutoptimization use for 2 dimensions runner.py: loads of optimization and changes (such as saving gtds, saving .csv files earlier etc.) tsNET.py: added gtds as parameter, changed hyperparameter (iterations) visualization_tool.py: changed a variable name

+++ runner.py
@@ -4,28 +4,76 @@ from glob import glob
 import numpy as np
 import pandas as pd
 import networkx as nx
+import time
+
 from sklearn import manifold
 from layout_techniques.tsNET import tsNET
 from layout_techniques.pivot_mds import pivot_mds
+from layout_techniques.fa2 import run_fa2
 
 import constants
 
 
 projection_names = {
-                    'A': ['SM', 'pivot_mds', 'tsNET', 'tsNETstar']
+                    'A': ['SM', 'pivot_mds', 'FA2', 'tsNET', 'tsNETstar']
                 }
 
+# projection_names = {
+#                     'A': ['SM', 'pivot_mds', 'FA2']
+#                 }
 
-def get_projections(graph, n_components):
 
-    layouts = {
-                    'SM': manifold.smacof(dissimilarities = nx.floyd_warshall_numpy(graph), n_components = n_components),
-                    'pivot_mds': pivot_mds(graph, dim = n_components),
-                    'tsNET' : tsNET(graph, star = False, dimensions = n_components),
-                    'tsNETstar' : tsNET(graph, star = True, dimensions = n_components)
-    }
+def get_projections(graph, n_components, gtds, technique = 'all'):
+
+    layouts = {}
+
+    if technique == 'SM' or technique == 'all':
+        start = time.time()
+        layouts['SM'] = manifold.smacof(dissimilarities = gtds, n_components = n_components)
+        print('SM took: ' + str(time.time() - start) + ' seconds')
+
+    if technique == 'pivot_mds' or technique == 'all':
+        start = time.time()
+        layouts['pivot_mds'] = pivot_mds(graph, dim = n_components, D = gtds)
+        print('pivot_mds took: ' + str(time.time() - start) + ' seconds')
+
+    if technique == 'FA2' or technique == 'all':
+        start = time.time()
+        bh_optimize = False
+        if n_components == 2:
+            bh_optimize = True
+        layouts['FA2'] = run_fa2.run_fa2(G = graph, pos = None, edge_weight = 0.0, max_iter = 100, dim = n_components, bh_optimize = bh_optimize)
+        print('FA2 took: ' + str(time.time() - start) + ' seconds')
+
+    if technique == 'tsNET' or technique == 'all':
+        try:
+            start = time.time()
+            layouts['tsNET'] = tsNET(graph, star = False, dimensions = n_components, gtds = gtds, n = 500)
+            print('tsNET took: ' + str(time.time() - start) + ' seconds')
+        except Exception as e:
+            print(e)
+            pass
+
+    if technique == 'tsNETstar' or technique == 'all':
+        try:
+            start = time.time()
+            layouts['tsNETstar'] = tsNET(graph, star = True, dimensions = n_components, gtds = gtds, n = 500),
+            print('tsNETstar took: ' + str(time.time() - start) + ' seconds')
+        except Exception as e:
+            print(e)
+            pass
 
     return layouts
+
+
+def save_gtds(graph, dataset_name):
+
+    gtds = nx.floyd_warshall_numpy(graph)
+    df = pd.DataFrame(gtds)
+
+    name = 'data/{0}/{0}-gtds.csv'.format(dataset_name)
+
+    df.to_csv(name, index=None, sep=';')
 
 
 if __name__ == '__main__':
@@ -38,8 +86,12 @@ if __name__ == '__main__':
     # layout technique selection, see above
     selection = 'A'
 
+    # if we want to overwrite the exact same files, e.g. grid-src.csv already exists, we don't need to visualize it again
+    overwrite = False
+
     for dataset_name in all_dataset_names:
 
+        print('Working on ' + str(dataset_name))
         input_file = glob('data/{0}/*-src.csv'.format(dataset_name))[0]
 
         df = pd.read_csv(input_file, sep=';', header=0)
@@ -48,40 +100,57 @@ if __name__ == '__main__':
 
         selected_projections = projection_names[selection]
 
+        gtds_file = 'data/{0}/{0}-gtds.csv'.format(dataset_name)
+        if not os.path.isfile(gtds_file):
+            print('Creating and saving the gtds')
+            # save the graph theoretic shortest distances
+            save_gtds(graph, dataset_name)
+
+        gtds = pd.read_csv(gtds_file, sep=';', header=0)
+        gtds = gtds.to_numpy()
+
         header = ['x', 'y', 'z']
 
         for n_components in [2, 3]:
-            projections = get_projections(graph, n_components)
 
             for proj_name in selected_projections:
 
-                if proj_name == 'SM':
-                    p, _ = projections[proj_name]
-                else:
-                    p = projections[proj_name]
-
                 output_file = os.path.join(constants.output_dir, '{0}-{1}-{2}d.csv'.format(dataset_name, proj_name, n_components))
 
-                print('dim: {0}, proj: {1}'.format(n_components, proj_name))
-                print('output_file: {0}'.format(output_file))
+                # if it doesnt exist
+                if not os.path.isfile(output_file) or overwrite:
+                    print('Doing ' + str(proj_name))
+                    projections = get_projections(graph, n_components, gtds, technique = proj_name)
 
-                if n_components == 2:
-                    # take the smallest value seen
-                    smallest_val = abs(min(np.min(p[:, 0]), np.min(p[:, 1])))
-                    # take the maximum of the the maximum differences (max difference of x coordinates & max difference of y coordinates)
-                    scale_factor = max(np.max(p[:, 0]) - np.min(p[:, 0]), np.max(p[:, 1]) - np.min(p[:, 1]))
+                    if proj_name == 'SM':
+                        p, _ = projections[proj_name]
+                    else:
+                        p = projections[proj_name]
+
+
+
+                    print('dim: {0}, proj: {1}'.format(n_components, proj_name))
+                    print('output_file: {0}'.format(output_file))
+
+                    if n_components == 2:
+                        # take the smallest value seen
+                        smallest_val = abs(min(np.min(p[:, 0]), np.min(p[:, 1])))
+                        # take the maximum of the the maximum differences (max difference of x coordinates & max difference of y coordinates)
+                        scale_factor = max(np.max(p[:, 0]) - np.min(p[:, 0]), np.max(p[:, 1]) - np.min(p[:, 1]))
+                    else:
+                        smallest_val = abs(min(np.min(p[:, 0]), np.min(p[:, 1]), np.min(p[:, 2])))
+                        scale_factor = max(np.max(p[:, 0]) - np.min(p[:, 0]), np.max(p[:, 1]) - np.min(p[:, 1]),
+                                           np.max(p[:, 2]) - np.min(p[:, 2]))
+
+                    # add the smallest value to the coordinates so it translates to coordinates >=0
+                    p += smallest_val
+
+                    # scale the layout
+                    coords = p / scale_factor
+
+                    df_layout = pd.DataFrame(coords)
+                    df_layout.columns = header[:n_components]
+
+                    df_layout.to_csv(output_file, index = None, sep = ';')
                 else:
-                    smallest_val = abs(min(np.min(p[:, 0]), np.min(p[:, 1]), np.min(p[:, 2])))
-                    scale_factor = max(np.max(p[:, 0]) - np.min(p[:, 0]), np.max(p[:, 1]) - np.min(p[:, 1]),
-                                       np.max(p[:, 2]) - np.min(p[:, 2]))
-
-                # add the smallest value to the coordinates so it translates to coordinates >=0
-                p += smallest_val
-
-                # scale the layout
-                coords = p / scale_factor
-
-                df_layout = pd.DataFrame(coords)
-                df_layout.columns = header[:n_components]
-
-                df_layout.to_csv(output_file, index = None, sep = ';')
+                    print(output_file + ' already exists')
